optional volume creation

+++ internal/cloud/digitalocean/delete.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/peterhalasz/envoi/internal/cloud"
 	log "github.com/sirupsen/logrus"
+	"github.com/spf13/viper"
 )
 
 func (p *DigitalOceanProvider) DeleteWorkstation(params *cloud.WorkstationDeleteParams) error {
@@ -26,20 +27,22 @@ func (p *DigitalOceanProvider) DeleteWorkstation(params *cloud.WorkstationDelete
 		return fmt.Errorf("workstation can't be deleted until at least 5 minutes old. Current age: %d minutes", workstation_age_minutes)
 	}
 
-	log.Debugf("Detaching workstation %d from volume %s", status.ID, status.Volume)
-	_, _, err := p.client.StorageActions.DetachByDropletID(context.TODO(), status.Volume, status.ID)
+	if viper.GetBool("digitalocean.volumes.enabled") {
+		log.Debugf("Detaching workstation %d from volume %s", status.ID, status.Volume)
+		_, _, err := p.client.StorageActions.DetachByDropletID(context.TODO(), status.Volume, status.ID)
 
-	if err != nil {
-		log.Debugf("Error %s", err.Error())
-		return err
+		if err != nil {
+			log.Debugf("Error %s", err.Error())
+			return err
+		}
+		log.Debugf("Workstation %d detached from volume %s", status.ID, status.Volume)
+
+		log.Debugf("Sleeping for 5 seconds")
+		time.Sleep(5 * time.Second)
 	}
-	log.Debugf("Workstation %d detached from volume %s", status.ID, status.Volume)
-
-	log.Debugf("Sleeping for 5 seconds")
-	time.Sleep(5 * time.Second)
 
 	log.Debugf("Deleting workstation %d", status.ID)
-	_, err = p.client.Droplets.Delete(context.TODO(), status.ID)
+	_, err := p.client.Droplets.Delete(context.TODO(), status.ID)
 
 	if err != nil {
 		log.Debugf("Error %s", err.Error())
@@ -47,16 +50,18 @@ func (p *DigitalOceanProvider) DeleteWorkstation(params *cloud.WorkstationDelete
 	}
 	log.Debugf("Workstation %d deleted", status.ID)
 
-	log.Debugf("Sleeping for 5 seconds")
-	time.Sleep(5 * time.Second)
+	if viper.GetBool("digitalocean.volumes.enabled") {
+		log.Debugf("Sleeping for 5 seconds")
+		time.Sleep(5 * time.Second)
 
-	log.Debugf("Deleting volume %s", status.Volume)
-	_, err = p.client.Storage.DeleteVolume(context.TODO(), status.Volume)
-	if err != nil {
-		log.Debugf("Error %s", err.Error())
-		return err
+		log.Debugf("Deleting volume %s", status.Volume)
+		_, err = p.client.Storage.DeleteVolume(context.TODO(), status.Volume)
+		if err != nil {
+			log.Debugf("Error %s", err.Error())
+			return err
+		}
+		log.Debugf("Volume %d deleted", status.ID)
 	}
-	log.Debugf("Volume %d deleted", status.ID)
 
 	return err
 }
